backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
apscheduler==3.10.4
anthropic==0.28.0
reportlab==4.1.0
openpyxl==3.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
